backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
python-dotenv==1.0.1
httpx==0.27.2
PyPDF2==3.0.1
python-docx==1.1.2
beautifulsoup4==4.12.3
requests==2.32.3
yt-dlp==2024.12.6
openai==1.55.3
pydub==0.25.1
